Some final tweaks

+++ impl/gamelib/state_menu.hpp
@@ -17,7 +17,7 @@ class StateMenu : public jt::GameState {
 public:
     std::string getName() const override;
 
-    void setScore(int s) { m_score = s; };
+    void setScore(float s) { m_score = s; };
 
 private:
     std::shared_ptr<jt::Shape> m_background;
@@ -30,7 +30,7 @@ private:
     std::shared_ptr<jt::Shape> m_overlay;
     std::shared_ptr<jt::Vignette> m_vignette;
 
-    int m_score { 0 };
+    float m_score { 0.0f };
 
     bool m_started { false };
